READ
This is the source code so far.  IMPORTANT: for this to be useful, create a new Android Studio Project and do "blank activity".  From there you can look at these and copy and past them. Reason is, there are other folders like "build" "gradle" etc. that bloat the folder. when creating the project, android studio will automatically make these anyway, and i didn't change the defaults, so it would be a waste.

ALSO: I only did anything to XML and .java files.  Other stuff like colors and styles are default.